Use base64 encoding for binary rng state information in textmode .sv files, and prefix with a comment '#'.
Saving binary data to a textfile is problematic, as it could confuse editors, parsers etc. The new file format is:

<blitz array>

#<base64 encoded rng state>~
# t dtTry

The '~' character is used as an additional safety check to ensure that the file format is correct and that the base64 string has the right length.
If it is not found where expected, an RNGStateParsingException is thrown.

Note that rfc4648 (base-64 encoding) requires padding of the binary data if its length is not an integral multiple of 3. Here we know the size of 
the encoded data, so we don't need to padd.