* implemented boost::serialization support for Randomized
Rationale: Storing additional data to the stream which an oarchive uses is discouraged, therefore we implement the necessary functions to support boost::serialization of Randomized objects directly. The save and load  functions are implemented in terms of the purely virtual readState and writeState functions, which have to be implemented by deriving classes anyway to support rng state restoration with text files. Now, when using binary output files, all relevant data (state vector, rng state, t, dtTry) can be written to and read from the archive directly.

Problem: *getRandomized() is of type const Randomized &, but we need Randomized & when restoring the state from the archive. At the moment a workaround is in place which uses const_cast<Randomized &>, but maybe an additional getter function which returns a mutable type would be a cleaner solution.


* in binary mode: save rank and dimension extents to the archive

Rationale: without this information it is not possible for PyCPPQed to read the binary blitz array. With this information available, boost::serialization can be used in the c++ extension module to restore the blitz array. The binary file format is now:
archive << RANK << d_1 << ... << d_RANK << psi << randomized << t << dtTry