Removed boost::serialization code from Randomized, instead the rng state is written directly to the stream in binary output (after the archive is written).
Randomized::SmartPtr is now shared_ptr to Randomized instead of const Randomized, functions that alter the internal rng state are not const anymore, i.e. everything involving sampling or restoring the state from stream.

Stream methods add an implementation ID to the stream on write and check the ID on read. Wrong IDs are rejected. RandomizedGSL uses the class name as ID. The ID must not contain whitespaces, ID and actual state data has to be separated by a whitespace, any number of white spaces are allowed before and after the ID.